requirements.txt
aiosmtplib==3.0.2
alabaster==0.7.16
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
cloudinary==1.44.1
colorama==0.4.6
coverage==7.9.1
Deprecated==1.2.18
dnspython==2.7.0
docutils==0.21.2
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.110.3
fastapi-mail==1.5.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
limits==4.2
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.6.0
pyasn1==0.6.1
pydantic==2.11.5
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.4
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==7.4.7
sphinx-autobuild==2024.10.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.41
starlette==0.37.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.27.1
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.23.0
